+++ src/pages/Transaction/Transaction.js
@@ -10,21 +10,11 @@ import { formatTime } from '../../common/helpers/formateTime';
 import { Card } from '../../components/Card/Card';
 import { HR } from '../../components/HR/HR';
 import { formatCurrency } from '../../common/helpers/formatCurrency';
-import { CategoryIcon } from '../../components/CategoryIcon/CategoryIcon';
-import styled from 'styled-components';
-
-export const IconTransactionContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin: 0 auto;
-  gap: 8px;
-`;
-
-export const StyledCategoryIcon = styled(CategoryIcon)`
-  margin: 0 0 8px 0;
-`;
+import {
+  IconTransactionContainer,
+  StyledCategoryIcon,
+} from './Transaction.style';
+import { LabelText } from './Transaction.style';
 
 export const Transaction = () => {
   const { transactionId, accountId } = useParams();
@@ -58,9 +48,14 @@ export const Transaction = () => {
         <Heading displayAs={2} size={1}>
           Transaction details
         </Heading>
-        <Text>Payment method {transaction.paymentMethod}</Text>
+        <LabelText>
+          <Text>Payment method</Text> <Text>{transaction.paymentMethod}</Text>
+        </LabelText>
         <HR />
-        <Text>Status {transaction.pending ? 'Pending' : 'Completed'}</Text>
+        <LabelText>
+          <Text>Status</Text>
+          <Text>{transaction.pending ? 'Pending' : 'Completed'}</Text>
+        </LabelText>
       </Card>
     </PageLayout>
   );
